e_shop/requirements.txt
Django==3.2
djangorestframework==3.13.1
django-filter==22.1
asgiref==3.5.2
pytz==2022.1
sqlparse==0.4.2
Markdown==3.3.7
Pillow==9.1.1
pytz==2022.1
drf-yasg==1.20.0
djoser==2.1.0
djangorestframework_simplejwt
django-phonenumber-field==6.3.0
django-phonenumbers==1.0.1
phonenumbers==8.12.51
psycopg2